Iot_Big_Project
In our project, we designed a door opening system with a defined card, however, the information measured with the distance sensor in front of the door is instantly sent to ThingSpeak with the esp8266 module we have installed on the Raspberry Pi Pico, and the distance information in front of the door (is there a vehicle or not?) is sent to the internet. appears over it. The RFID card tag is defined to the Raspberry Pi Pico. When the valid card is read, the engine starts (door opens), the buzzer sounds, the blue LED lights up and the distance information on the first line and “Card Detected” are written on the second line on the LCD screen. When an undefined card is read, the engine does not start (the door does not open), the red LED lights, and the buzzer sounds. On the LCD screen, since the card is invalid, distance information is not written and “Invalid Card” is written on the screen.